Conflict resolutions, each fusing the PR's Texture->Field rename with a
master-side behavior change:

- core/synced.h: master replaced the hand-written traverse_1_cb_ro/rw pair
  with DRJIT_TRAVERSE(field, m_value) (drjit traversal rework); the PR
  renamed the class field->synced, so the macro takes the new class name:
  DRJIT_TRAVERSE(synced, m_value).
- render/shape.h: master moved m_texture_attributes from tsl::robin_map to
  std::map with a transparent comparator; the PR retyped the values to
  ref<Field>. Take both: std::map<std::string, ref<Field>, std::less<>>.
- emitters/area.cpp: union of the PR's get_emissive_surface_field + spectral
  requirement and master's two-sided support (b9286db1): m_twosided member
  and its Properties read are kept alongside the Field-typed radiance.
- fields/gridvolume.cpp: master dropped the duplicated m_accel argument to
  the dr::Texture3f constructor (texture state-machine rework); the PR's
  adjacent per-channel-max reset statements are kept.
- render/texture.cpp: master's improved eval_1_grad error (Throw with
  class_name()) on the PR's SurfaceField spelling.
- python/__init__.py: both variant hooks run -- the PR's built-in Python
  field registration first, then master's entry-point plugin (re)import.
- render/tests/test_texture.py: the dummy carries both the PR's out_dim
  and master's __repr__.
- textures/math.cpp -> fields/math.cpp: master's new expression texture
  (a7699ce8 / PR #1944) moved into the PR's renamed directory and ported to
  the Field API: SurfaceField base, Field-typed inputs gated on
  supports_surface_queries(), out_type()/out_dim() propagating the joint
  input type (Float for a constant expression), MI_TRAVERSE_CB on
  SurfaceField. Its CMake target keeps master's math_tex/OUTPUT_NAME dance
  (Embree-internal "math" target clash), now in src/fields/CMakeLists.txt.
- src/textures/CMakeLists.txt, texture_v.cpp, volume_v.cpp: deletions kept
  (the PR replaces them with src/fields/ + field_v.cpp); master's changes to
  the deleted binding files (nanobind-3 trampoline updates) are ported to
  field_v.cpp in a follow-up commit.
- docstr.h: the PR-side deletion of the __doc_mitsuba_Texture_* and
  __doc_mitsuba_Volume_* blocks stands (their binding files are deleted);
  no surviving DOC() reference names them. Regenerate with
  resources/mkdocs.py once the stack builds.

+++ include/mitsuba/python/synced.h
@@ -19,7 +19,7 @@ struct type_caster<mitsuba::synced<DeviceType, HostType>> {
     template <typename T> using Cast = Value;
     static constexpr auto Name       = Caster::Name;
 
-    bool from_python(handle src, uint8_t flags,
+    bool from_python(handle src, uint32_t flags,
                      cleanup_list *cleanup) noexcept {
         return caster.from_python(src, flags, cleanup);
     }
